Changed the time format

+++ emergency caller/ContentView.swift
@@ -77,7 +77,7 @@ struct ContentView: View {
 private let itemFormatter: DateFormatter = {
     let formatter = DateFormatter()
     formatter.dateStyle = .short
-    formatter.timeStyle = .short
+    formatter.timeStyle = .long
     return formatter
 }()
